Restart from scratch, construct the project from a network perspective
It's not really a total rewrite because a lot of old code will be reused.
We have learned a lot since our first attempt and we will try to do better this time.

For now, we are focusing on the network aspect of the game, then we will build
little by little the rendering.  We aim to get at the same state as we were before
the rewrite.  With maybe better rendering technique and cleaner code.

Instead of GLFW, SDL2 will be used to create the OpenGL context and handle events.
The network code is still done with POSIX calls, despite the fact that SDL
provide some network abstraction.

Fresh new start :)