Make the client send reliably player diff to the server
The code need to be polished, but the model work.  I'm not going to
describe it precisely since I should write a document about the network
protocol soon.  The main ideas of the protocol is to use UDP and don't
wait for acknowledgement before sending the next chunk of data, like a
real-time protocol.

For now the server only handle one client, and the player diff are not
well tested.  Also the 'struct player_array' is not yet reusable for
'struct world', and is not yet ready for multi player.

The network protocol is inspired from the one of Quake 3, but I'm not sure
if I implemented the exact same protocol or if I came with a different
one.  Also in Quake 3 the server ended up by maintening the equivalent
structure 'player_array' for _each_ connections server-side, where in
mine I have an idea to have a (simple) shared array for every players
connected.

A packet contain 3 fields, and then a player diff (for now)
  - uint16_t seq_from, seq_to: ID from where the diff start and where it
    get to.  'seq_from' must be know by the client where seq_to can be a
    new one.
  - uint16_t ack: Acknowledgement of the last well received packet.  The
    acknowledgement must refer to a previous seq ID received.

	+---+---+---+---+---+---+=== ... ===+
	| from  |  to   |  ack  |   diff    |
	+---+---+---+---+---+---+=== ... ===+

I have let some debug printf for now because I know that some function
need to be reworked.