client: Randomly change player name each tick
Server now print the last player name known.  It appears that there is
many cases I didn't think about.  For example we should always take the
more recent player name, as it is, I think it doesn't.  It also show
that player_array is not really a good API right now.

Above 16 names sended, the server do not handle them coprrectly anymore,
because the player array is full and I didn't write the code to flush
it...