Move some network logic into player_array
The code is not as nice as I wanted: I feel I can do even better.  But
the code to manage the network is now far more readable.  Player array
still do not means much and I think a more accurate name may be
"snapshots list" or something like that.

Player array has now all it needs to be generic, at least in the case
with one client and one server.  It will not be suitable for the server
to maange world snapshots for example, but I guess it can be improved in
this sens.

The result of my test is a bit unexpected.  When the client send a new
snapshot, it will wait for the ack, and when it comes I expected him to
send a empty diff to notice the server that he has nothing to send.  But
It doesn't.  You might think: "WTF fix it!  You don't even understand
your own protocol?"  And it turns out I don't.  I must take some
time to think about it, and then write a clear document once I've
understand everything.

So for now the behavior of the client maybe the right one.  The old one
may also be correct, but a bit more complicated (why send empty diff
after all?).

Anyway, I will obviously fix all this mess.