server: Move network client management into a new file
Introduce a netclient structure, netclient stand for network client.
For now it only contains a addres, but in the futur it would conatins a
lot of other things.

Because of the reason explained above, I wonder if I should split this
struct in two parts : a netconn, for network connection and a netclient
struct who include the netconn struct.  We will see how it evolve.

This commit also add a netclient_list, wich is as you guessed a list of
netclient.  It then help to retrieve a netclient struct from his
addres.  It is needed because when receiving network data, the recv()
syscall only fill a network addres.